https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved emailBodyService to esn.inbox.libs
