+++ src/esn.inbox.libs/app/app.module.js
@@ -9,3 +9,4 @@ require('esn-frontend-common-libs/src/frontend/js/modules/http.js');
 require('./app.config');
 require('./app.constants');
 require('./services/inbox-restangular.service');
+require('./services/email-body/email-body');
