+++ src/linagora.esn.unifiedinbox/app/components/composer/body-editor/text/composer-body-editor-text.controller.js
@@ -1,4 +1,4 @@
-require('../../../../services/email-body/email-body.js');
+require('../../../../../../esn.inbox.libs/app/services/email-body/email-body.js');
 
 (function(angular) {
   'use strict';
