+++ src/linagora.esn.unifiedinbox/app/directives/main.js
@@ -3,7 +3,7 @@ require('../services/new-composer/new-composer.js');
 require('../services/custom-role-mailbox/custom-role-mailbox.service.js');
 require('../services/mailboxes/mailboxes-service.js');
 require('../services/jmap-item/jmap-item-service.js');
-require('../services/email-body/email-body.js');
+require('../../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('../services/shortcuts/shortcuts.constants.js');
 require('../services/filtering/filtering-service.js');
 require('../services/plugins/plugins.js');
