+++ src/linagora.esn.unifiedinbox/app/filters.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-require('./services/email-body/email-body.js');
+require('../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('./services/mailboxes/mailboxes-service.js');
 
 (function (angular) {
