+++ src/linagora.esn.unifiedinbox/app/services.js
@@ -6,7 +6,7 @@ require('./services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('./services/jmap-helper/jmap-helper.js');
 require('./services/mailboxes/mailboxes-service.js');
 require('./services/hook/email-sending-hook.service.js');
-require('./services/email-body/email-body.js');
+require('../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('./services/attachment-jmap/attachment-jmap.constants.js');
 require('./services/email-resolver/inbox-email-resolver.service.js');
 require('./services/jmap-client-provider/jmap-client-provider.js');
