+++ src/linagora.esn.unifiedinbox/app/services/draft/draft.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-require('../../services/email-body/email-body.js');
+require('../../../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('../../services.js');
 require('../../services/jmap-helper/jmap-helper.js');
 require('../../services/config/config.js');
